Move requires to hyperclient class to avoid circular require issues
The Resource class references LinkCollection (which references Link) and
ResourceCollection and both Link and ResourceCollection reference
Resource. Because each class required what it referenced, this created
a circular require cycle (which ruby works around, but also issues an
warning).

To fix this, this commit moves all requires to the `hyperclient.rb`, thus
making sure we only require files once and avoid any cycles.

Thanks @dblock for suggesting this approach.