Require Faraday >= 0.9.0 in gemspec
In #115 we started configuring Faraday to use the FlatParamsEncoder to
avoid an issue with dropped parameters when building a request url.

Unfortunately, as FlatParamsEncoder was only added on Faraday 0.9.0
we should require that version as a minimum, otherwise clients may get
confusing error messages after upgrading to the latest hyperclient
version.

Fixes #117