Fix link delegation returning nil for field with value false
Whenever a resource includes a field that is `false`, trying to access
it through a `Link` returned `nil`, rather than `false`.

This is because `Link` confused the `false` response with the resource
not actually handling the value, and then tried to use the delegation as
if a collection method was being called on the resource.

Fixes #126